+++ benchmarks/bench_chunk_delta_h.py
@@ -50,7 +50,7 @@ import torch
 from benchmarks.utils import benchmark_cuda_mode_fn, relative_rms_error_max_mean_abs
 
 # ─── CuTe DSL wrapper (TVM-FFI compile cache) ───
-_delta_h_mod = importlib.import_module("cula.ops.chunk_delta_h_sm100")
+_delta_h_mod = importlib.import_module("cula.ops.kda.sm100.delta_h")
 chunk_gated_delta_rule_fwd_h = _delta_h_mod.chunk_gated_delta_rule_fwd_h
 
 # ─── FLA baseline imports ───
