+++ benchmarks/bench_fwd_o.py
@@ -47,7 +47,7 @@ import torch
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parent.parent))
 
 # ─── CuTe DSL wrapper (TVM-FFI compile cache) ───
-_fwd_o_mod = importlib.import_module("cula.ops.fwd_o_sm100")
+_fwd_o_mod = importlib.import_module("cula.ops.kda.sm100.fwd_o")
 chunk_gla_fwd_o = _fwd_o_mod.chunk_gla_fwd_o
 build_chunk_indices = _fwd_o_mod.build_chunk_indices
