+++ cula/kda/chunk_bwd.py
@@ -14,6 +14,8 @@
 
 # Copyright (c) 2023-2025, Songlin Yang, Yu Zhang
 
+"""SM100 modular chunk KDA backward orchestration"""
+
 import importlib
 
 import torch
@@ -37,10 +39,10 @@ from fla.utils import (
 
 import cula.cudac as cula_cuda
 from cula.kda.chunk_intra import chunk_kda_bwd_intra
-from cula.ops.chunk_wy_dqkg_sm100 import chunk_kda_bwd_wy_dqkg_fused as chunk_kda_bwd_wy_dqkg_fused_cutedsl
+from cula.ops.kda.sm100.bwd_wy_dqkg import chunk_kda_bwd_wy_dqkg_fused as chunk_kda_bwd_wy_dqkg_fused_cutedsl
 from cula.utils import prepare_uniform_cu_seqlens
 
-_delta_h_mod = importlib.import_module("cula.ops.chunk_delta_h_sm100")
+_delta_h_mod = importlib.import_module("cula.ops.kda.sm100.delta_h")
 chunk_gated_delta_rule_fwd_h = _delta_h_mod.chunk_gated_delta_rule_fwd_h
 
 BK_LIST = [32, 64] if check_shared_mem() else [16, 32]
