--- cula/ops/cp/__init__.py
@@ -1,3 +1,0 @@
-from cula.ops.cp.chunk_delta_h import intracard_fwd_h
-
-__all__ = ["intracard_fwd_h"]
