+++ cula/ops/kda/experimental/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2025-2026 Ant Group Co., Ltd.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Experimental / unwired KDA kernels — not part of any production path."""
