+++ cula/ops/kda/sm100/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2025-2026 Ant Group Co., Ltd.
+# SPDX-License-Identifier: Apache-2.0
+
+"""SM100 (Blackwell) modular-chunk KDA backend kernels."""
