+++ cula/ops/kda/sm100/bwd_wy_dqkg.py
@@ -24,19 +24,15 @@ from cutlass.cute.tensor import TensorSSA
 from cutlass.cute.typing import BFloat16, Float32, Int32, Int64
 from fla.ops.utils import prepare_chunk_indices
 
-from cula.ops.intrinsics_sm100 import (
-    reinterpret_cast,
-    store_256b,
-    subvec,
+from cula.ops.ptx import reinterpret_cast, store_256b, subvec
+from cula.ops.sm100.ptx import (
+    Tcgen05SmemDescriptor,
     tcgen05_fence_after,
     tcgen05_fence_before,
     tcgen05_ld_32x32b,
     tcgen05_st_32x32b,
-    umma_arrive,
-)
-from cula.ops.ptx_umma_ext import (
-    Tcgen05SmemDescriptor,
     tcgen05mma_ws_ss_f16,
+    umma_arrive,
 )
 from cula.utils import USE_FAST_MATH, assert_blackwell, prepare_uniform_cu_seqlens
