+++ cula/ops/kda/sm100/cp/__init__.py
@@ -1,0 +1,8 @@
+# Copyright 2025-2026 Ant Group Co., Ltd.
+# SPDX-License-Identifier: Apache-2.0
+
+"""SM100 intracard context-parallel backend for the chunk delta-rule recurrence."""
+
+from cula.ops.kda.sm100.cp.chunk_delta_h import intracard_fwd_h
+
+__all__ = ["intracard_fwd_h"]
