+++ cula/ops/kda/sm100/cp/chunk_delta_h.py
@@ -26,7 +26,7 @@ Pipeline (3 stages):
 Reference:
     - FLA intra-card CP: fla/ops/common/intracard_cp.py
     - FLA CP kernels:    fla/ops/cp/chunk_delta_h.py
-    - cuLA chunk_delta_h: cula/ops/chunk_delta_h.py
+    - cuLA chunk_delta_h: cula/ops/kda/sm100/delta_h.py
 """
 
 from __future__ import annotations
@@ -40,14 +40,14 @@ import torch
 
 from cula.utils import get_device_sm_count, get_pre_scan
 
-# Lazy import to avoid circular dependency with cula.ops.chunk_delta_h
+# Lazy import to avoid circular dependency with cula.ops.kda.sm100.delta_h
 _chunk_gated_delta_rule_fwd_h = None
 
 
 def _get_fwd_h():
     global _chunk_gated_delta_rule_fwd_h
     if _chunk_gated_delta_rule_fwd_h is None:
-        from cula.ops.chunk_delta_h_sm100 import chunk_gated_delta_rule_fwd_h
+        from cula.ops.kda.sm100.delta_h import chunk_gated_delta_rule_fwd_h
 
         _chunk_gated_delta_rule_fwd_h = chunk_gated_delta_rule_fwd_h
     return _chunk_gated_delta_rule_fwd_h
@@ -347,12 +347,12 @@ def intracard_merge(
     For split seq [s0, s1, ..., s_{n-1}]: h0_sj = m_{j-1} @ h0_{j-1} + he_{j-1}.
     Returns (initial_states_merge [num_non_first, H, K, V] fp32, num_non_first).
     """
-    from cula.ops.cp.merge import merge_fwd
+    from cula.ops.kda.sm100.cp.merge import launch_merge
 
     if num_non_first == 0:
         return None, 0
 
-    initial_states_merge = merge_fwd(
+    initial_states_merge = launch_merge(
         hm=hm,
         seq_starts=merge_seq_starts,
         seq_counts=merge_seq_counts,
@@ -413,6 +413,8 @@ def intracard_fwd_h(
     cu_seqlens_cpu: torch.Tensor | None = None,
     chunk_indices: torch.Tensor | None = None,
     max_splits: int = 32,
+    allow_fallback: bool = True,
+    skip_precheck: bool = False,
 ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor | None]:
     """Intra-card CP chunk_delta_h forward; drop-in replacement for chunk_gated_delta_rule_fwd_h.
 
@@ -429,7 +431,9 @@ def intracard_fwd_h(
     if cu_seqlens_cpu is None:
         cu_seqlens_cpu = cu_seqlens.cpu()
 
-    if not should_use_intracard_cp(cu_seqlens_cpu, num_sms, H, chunk_size):
+    if not skip_precheck and not should_use_intracard_cp(cu_seqlens_cpu, num_sms, H, chunk_size):
+        if not allow_fallback:
+            raise ValueError("SM100 intracard CP is rejected by the pre-split dispatch policy.")
         return _get_fwd_h()(
             k=k,
             w=w,
@@ -473,6 +477,8 @@ def intracard_fwd_h(
         split_info = False
 
     if not split_info:
+        if not allow_fallback:
+            raise ValueError("SM100 intracard CP is not meaningfully splittable for this shape.")
         return _get_fwd_h()(
             k=k,
             w=w,
