+++ tests/test_la_decode.py
@@ -30,7 +30,7 @@ import torch
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parent.parent))
 
 
-from cula.ops.la_decode import linear_attention_decode
+from cula.ops.lightning.decode import linear_attention_decode
 
 try:
     from fla.ops.common.fused_recurrent import fused_recurrent_fwd
@@ -231,9 +231,10 @@ def test_vs_fla(B):
 
 # End-to-End Prefill -> Decode Test
 # ---------------------------------------------------------------------------
+@pytest.mark.sm100_only
 def test_prefill_decode_e2e():
     """Verify prefill output state passes directly into decode without transpose."""
-    from cula.ops.lightning_attn_sm100 import lightning_attn_fwd
+    from cula.ops.lightning.prefill_sm100 import lightning_attn_fwd
 
     B, S, H, D = 2, 64, 8, 128
     scale = D**-0.5
